requirements.txt
ConfigArgParse==0.11.0
Flask==1.1.1
Flask-Cors==3.0.8
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==4.1.0
Werkzeug==1.0.0
click==7.1.1
flask-csrf==0.9.2
gunicorn==20.0.4
hiredis==1.0.1
itsdangerous==1.1.0
redis==3.4.1
redis-lru==0.1.0
redisearch==0.9.0
rmtest==0.7.0
six==1.14.0
virtualenv==15.1.0
